hardware/requirements.txt
websockets
Pillow
httpx

# Raspberry Pi specific (uncomment when deploying to Pi)
# RPi.GPIO
# waveshare-epd
# luma.oled
